ex2019_2_imersaoii-master/requirements.txt
colorama==0.4.1

dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-qr-code==1.0.0
gunicorn==20.0.0
Pillow==6.2.1
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
qrcode==6.1
six==1.12.0
sqlparse==0.3.0
virtualenv==16.7.3
virtualenvwrapper-win==1.2.5
whitenoise==4.1.4

